changed backend to backend-stub, removed unused test code in backend/index.js
