+++ .github/workflows/dev_ctofeedback.yml
@@ -14,7 +14,7 @@ on:
   workflow_dispatch:
 
 env:
-  AZURE_FUNCTIONAPP_PACKAGE_PATH: "./backend-stub" # set this to the path to your web app project, defaults to the repository root
+  AZURE_FUNCTIONAPP_PACKAGE_PATH: "./backend" # set this to the path to your web app project, defaults to the repository root
   NODE_VERSION: "14.x" # set this to the node version to use (supports 8.x, 10.x, 12.x)
 
 jobs:
