+++ .vscode/tasks.json
@@ -8,7 +8,7 @@
       "isBackground": true,
       "dependsOn": "npm install (functions)",
       "options": {
-        "cwd": "${workspaceFolder}/backend-stub"
+        "cwd": "${workspaceFolder}/backend"
       }
     },
     {
@@ -16,7 +16,7 @@
       "label": "npm install (functions)",
       "command": "npm install",
       "options": {
-        "cwd": "${workspaceFolder}/backend-stub"
+        "cwd": "${workspaceFolder}/backend"
       }
     },
     {
@@ -25,7 +25,7 @@
       "command": "npm prune --production",
       "problemMatcher": [],
       "options": {
-        "cwd": "${workspaceFolder}/backend-stub"
+        "cwd": "${workspaceFolder}/backend"
       }
     }
   ]
